Move population task to seed script
